requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.3.0
google-generativeai==0.3.2
python-dotenv==1.0.0
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
requests==2.31.0
pydantic==2.5.0 